Add Fabric 1.19.4 and update Gradle to 8.1 since Fabric requires it (Thus removed forge versions)
