+++ fabric-1.19.4/gradle.properties
@@ -1,0 +1,17 @@
+# Done to increase the memory available to gradle.
+org.gradle.jvmargs=-Xmx1G
+org.gradle.parallel=true
+
+# Fabric Properties
+# check these on https://fabricmc.net/develop
+minecraft_version=1.19.4
+yarn_mappings=1.19.4+build.2
+loader_version=0.14.19
+
+# Mod Properties
+mod_version = 1.0.0
+maven_group = com.example
+archives_base_name = fabric-example-mod
+
+# Dependencies
+fabric_version=0.79.0+1.19.4
