+++ fabric-1.19.4/src/main/resources/assets/mpkmod/lang/pl_pl.json
@@ -1,0 +1,8 @@
+{
+  "mpkmod.key.main_gui.desc": "Otwórz menu MPK",
+  "mpkmod.key.lb_gui.desc": "Otwórz menu Landing Block",
+  "mpkmod.key.lb_set.desc": "Ustaw Landing Block",
+  "mpkmod.key.options_gui.desc": "Otwórz menu ustawień",
+  "mpkmod.main_gui.title": "Menu MPK",
+  "mpkmod.lb_gui.title": "Menu Landing Block"
+}
