--- forge-1.14.4/src/main/resources/assets/mpkmod/lang/pl_pl.lang
@@ -1,4 +1,0 @@
-mpkmod.key.main_gui.desc=Otwórz menu Moda MPK
-mpkmod.key.lb_gui.desc=Otwórz menu bloku lądowania
-mpkmod.main_gui.title=Menu Moda MPK
-mpkmod.lb_gui.title=Menu bloku lądowania
