--- forge-1.19.4/build.gradle
@@ -1,56 +1,0 @@
-plugins {
-    id 'eclipse'
-    id 'net.minecraftforge.gradle' version '5.1.+'
-}
-
-java.toolchain.languageVersion = JavaLanguageVersion.of(17)
-
-tasks.withType(JavaCompile) {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-}
-
-println "Java: ${System.getProperty 'java.version'}, JVM: ${System.getProperty 'java.vm.version'} (${System.getProperty 'java.vendor'}), Arch: ${System.getProperty 'os.arch'}"
-minecraft {
-    mappings channel: 'official', version: '1.19.4'
-
-    runs {
-        client {
-            workingDirectory project.file('../run')
-
-            // Recommended logging data for a userdev environment
-            // The markers can be added/remove as needed separated by commas.
-            // "SCAN": For mods scan.
-            // "REGISTRIES": For firing of registry events.
-            // "REGISTRYDUMP": For getting the contents of all registries.
-            property 'forge.logging.markers', 'REGISTRIES'
-
-            // Recommended logging level for the console
-            // You can set various levels here.
-            // Please read: https://stackoverflow.com/questions/2031163/when-to-use-the-different-log-levels
-            property 'forge.logging.console.level', 'debug'
-
-            // Comma-separated list of namespaces to load gametests from. Empty = all namespaces.
-            property 'forge.enabledGameTestNamespaces', 'mpkmod'
-
-            mods {
-                mpkmod {
-                    source sourceSets.main
-                }
-            }
-        }
-    }
-}
-
-// you have to rerun genIntellijRuns every time you change this
-dependencies {
-    minecraft 'net.minecraftforge:forge:1.19.4-45.0.55'
-    library project(":common")
-}
-
-//fix ClassNotFoundException during runClient because of dependencies not being added to classpath in dev environment
-minecraft.runs.all {
-    lazyToken('minecraft_classpath') {
-        configurations.library.copyRecursive().resolve().collect { it.absolutePath }.join(File.pathSeparator)
-    }
-}
