--- forge-1.19.4/gradle.properties
@@ -1,1 +1,0 @@
-$mcVersion=1.19.4
