--- forge-1.19.4/src/main/resources/META-INF/mods.toml
@@ -1,53 +1,0 @@
-# This is an example mods.toml file. It contains the data relating to the loading mods.
-# There are several mandatory fields (#mandatory), and many more that are optional (#optional).
-# The overall format is standard TOML format, v0.5.0.
-# Note that there are a couple of TOML lists in this file.
-# Find more information on toml format here:  https://github.com/toml-lang/toml
-# The name of the mod loader type to load - for regular FML @Mod mods it should be javafml
-modLoader="javafml" #mandatory
-# A version range to match for said mod loader - for regular FML @Mod it will be the forge version
-loaderVersion="[11,)" #mandatory (26 is current forge version)
-# The license for you mod. This is mandatory metadata and allows for easier comprehension of your redistributive properties.
-# Review your options at https://choosealicense.com/. All rights reserved is the default copyright stance, and is thus the default here.
-license="All rights reserved"
-# A URL to refer people to when problems occur with this mod
-issueTrackerURL="http://my.issue.tracker/" #optional
-# A list of mods - how many allowed here is determined by the individual mod loader
-[[mods]] #mandatory
-# The modid of the mod
-modId="mpkmod" #mandatory
-# The version number of the mod - there's a few well known ${} variables useable here or just hardcode it
-version="${file.jarVersion}" #mandatory
-# A display name for the mod
-displayName="MPK Mod" #mandatory
-# A URL to query for updates for this mod. See the JSON update specification <here>
-#updateJSONURL="https://raw.githubusercontent.com/kurrycat2004/MpkMod/main/update.json" #optional
-# A URL for the "homepage" for this mod, displayed in the mod UI
-#displayURL="http://example.com/" #optional
-# A file name (in the root of the mod JAR) containing a logo for display
-logoFile= "../mpkmod_logo.png" #optional
-# A text field displayed in the mod UI
-credits="" #optional
-# A text field displayed in the mod UI
-authors="" #optional
-# The description text for the mod (multi line!) (#mandatory)
-description='''
-
-
-A Mod that adds some pk features to the game
-'''
-# A dependency - use the . to indicate dependency for a specific modid. Dependencies are optional.
-
-#[[dependencies.mpkmod]] #optional
-#modId="forge"
-#mandatory=true
-#versionRange="[11,)"
-#ordering="NONE"
-#side="BOTH"
-
-#[[dependencies.mpkmod]]
-#modId="minecraft"
-#mandatory=true
-#versionRange="[1.8.9,)"
-#ordering="NONE"
-#side="BOTH"
