--- forge-1.19.4/src/main/resources/assets/mpkmod/lang/pl_pl.json
@@ -1,6 +1,0 @@
-{
-  "mpkmod.key.main_gui.desc": "Otwórz menu Moda MPK",
-  "mpkmod.key.lb_gui.desc": "Otwórz menu bloku lądowania",
-  "mpkmod.main_gui.title": "Menu Moda MPK",
-  "mpkmod.lb_gui.title": "Menu bloku lądowania"
-}
