--- forge-1.19.4/src/main/resources/pack.mcmeta
@@ -1,7 +1,0 @@
-{
-    "pack": {
-        "description": "mpkmod resources",
-        "pack_format": 4,
-        "_comment": "A pack_format of 4 requires json lang files. Note: we require v4 pack meta for all mods."
-    }
-}
