--- forge-1.8.9/gradle.properties
@@ -1,1 +1,0 @@
-$mcVersion=1.8.9
