+++ settings.gradle
@@ -5,10 +5,6 @@ pluginManagement {
             url = 'https://maven.wagyourtail.xyz/snapshots'
         }
         maven {
-            name = 'ForgeMaven'
-            url = 'https://maven.minecraftforge.net/'
-        }
-        maven {
             name = 'FabricMaven'
             url = 'https://maven.fabricmc.net/'
         }
@@ -18,6 +14,4 @@ pluginManagement {
 }
 
 include 'common'
-include 'forge-1.8.9'
-//include 'forge-1.14.4' // - not working
-include 'forge-1.19.4'
+include 'fabric-1.19.4'
